+++ source/versions.js
@@ -6,8 +6,8 @@ const { allNodeVersions } = require('./data')
 // Works for 1, 1.1, or 1.1.1, or 1.1.1 - sadasd
 function versionComparator(a, b) {
 	// https://github.com/substack/versionComparator-compare/pull/4
-	const pa = a.split('.')
-	const pb = b.split('.')
+	const pa = String(a).split('.')
+	const pb = String(b).split('.')
 	for (let i = 0; i < Math.min(pa.length, pb.length); i++) {
 		const na = Number(pa[i])
 		const nb = Number(pb[i])